+++ src/main/java/com/goorm/clonestagram/file/dto/update/VideoUpdateResDto.java
@@ -1,19 +1,15 @@
-package com.goorm.clonestagram.file.dto;
+package com.goorm.clonestagram.file.dto.update;
 
 import com.goorm.clonestagram.file.ContentType;
 import lombok.Builder;
 import lombok.Getter;
 
 import java.time.LocalDateTime;
-
-/**
- * 이미지 업로드 응답을 위한 DTO
- * - content, type, createdAt을 반환
- */
 @Getter
 @Builder
-public class ImageUploadResDto {
+public class VideoUpdateResDto {
     private String content;
     private ContentType type;
-    private LocalDateTime createdAt;
+    private LocalDateTime updatedAt;
 }
+
